refactor: move HighlightLabel.kt to components package
